+++ less1.1/src/App.js
@@ -1,30 +1,14 @@
 import React from 'react';
-
-let todoList=[
-  {
-    id: 1,
-    title: "Go shopping",
-  }, 
-  {
-    id: 2,
-    title: "Relax",
-  },
-  {
-    id: 3,
-    title: "Cook dinner",
-  }
-];
+import TodoList from './TodoList.js';
+import AddTodoForm from './AddTodoForm.js';
 
 function App() {
   return (
     // <> called fragment, used instead of div not to create to many divs
     <> 
       <h1>TODO list</h1>
-      <ul>
-        {todoList.map(function(item) {
-          return <li key={item.id}>{item.title}</li>;
-        })}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </>
   );
 }
